requirements.txt
TTS==0.22.0
torch==2.1.2
torchaudio==2.1.2
moviepy==1.0.3
pydub==0.25.1
PyYAML==6.0.1
numpy==1.22.0
scipy>=1.11.2
soundfile==0.12.1
